requirements.txt
# UI and Database
PySide6>=6.5.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
qt-material>=2.14

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.2

# Reporting and QR Code
reportlab>=4.0.4
qrcode>=7.4.2
pillow>=9.5.0

# Security and System
passlib>=1.7.4
psutil>=5.9.5

# Build Tools
pyinstaller>=5.10.1
PyUpdater>=4.0; python_version < "3.12"

# Optional Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2